Revert "Multi Tenancy"
