--- src/main/java/com/studies/ecommerce/controller/EcmController.java
@@ -1,13 +1,0 @@
-package com.studies.ecommerce.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class EcmController {
-
-    @GetMapping
-    public String index() {
-        return "index";
-    }
-}
