--- src/main/java/com/studies/ecommerce/hibernate/EcmMachineMultiTenantConnectionProvider.java
@@ -1,110 +1,0 @@
-package com.studies.ecommerce.hibernate;
-
-import org.hibernate.engine.config.spi.ConfigurationService;
-import org.hibernate.engine.jdbc.connections.spi.ConnectionProvider;
-import org.hibernate.engine.jdbc.connections.spi.MultiTenantConnectionProvider;
-import org.hibernate.hikaricp.internal.HikariCPConnectionProvider;
-import org.hibernate.service.spi.ServiceRegistryAwareService;
-import org.hibernate.service.spi.ServiceRegistryImplementor;
-import org.hibernate.service.spi.Startable;
-
-import java.sql.Connection;
-import java.sql.SQLException;
-import java.util.HashMap;
-import java.util.Map;
-
-public class EcmMachineMultiTenantConnectionProvider implements
-        MultiTenantConnectionProvider, ServiceRegistryAwareService, Startable {
-
-    private Map<String, ConnectionProvider> connectionProviders = null;
-
-    private Map properties = null;
-
-    @Override
-    public Connection getConnection(String tenantIdentifier) throws SQLException {
-        Connection connection = connectionProviders
-                .get(tenantIdentifier)
-                .getConnection();
-
-        return connection;
-    }
-
-    @Override
-    public Connection getAnyConnection() throws SQLException {
-        return getAnyConnectionProvider()
-                .getConnection();
-    }
-
-    @Override
-    public void releaseConnection(String tenantIdentifier, Connection connection) throws SQLException {
-        releaseAnyConnection(connection);
-    }
-
-    @Override
-    public void releaseAnyConnection(Connection connection) throws SQLException {
-        getAnyConnectionProvider().closeConnection(connection);
-    }
-
-    @Override
-    public boolean supportsAggressiveRelease() {
-        return getAnyConnectionProvider().supportsAggressiveRelease();
-    }
-
-    @Override
-    public boolean isUnwrappableAs(Class unwrapType) {
-        return getAnyConnectionProvider().isUnwrappableAs(unwrapType);
-    }
-
-    @Override
-    public <T> T unwrap(Class<T> unwrapType) {
-        return getAnyConnectionProvider().unwrap(unwrapType);
-    }
-
-    @Override
-    public void injectServices(ServiceRegistryImplementor serviceRegistry) {
-        this.properties = serviceRegistry
-                .getService(ConfigurationService.class)
-                .getSettings();
-    }
-
-    @Override
-    public void start() {
-        connectionProviders = new HashMap<>();
-
-        setTenant("host1_ecommerce",
-                "jdbc:mysql://localhost/db1_ecommerce?createDatabaseIfNotExist=true&useTimezone=true&serverTimezone=UTC",
-                "root",
-                "123456");
-
-        setTenant("host2_ecommerce",
-                "jdbc:mysql://localhost/db2_ecommerce?createDatabaseIfNotExist=true&useTimezone=true&serverTimezone=UTC",
-                "root",
-                "123456");
-
-        this.properties = null;
-    }
-
-    private void setTenant(String tenant, String url, String user, String password) {
-        Map<String, String> properties = new HashMap<>(this.properties);
-
-        properties.put("javax.persistence.jdbc.url", url);
-        properties.put("hibernate.connection.url", url);
-
-        properties.put("javax.persistence.jdbc.user", user);
-        properties.put("hibernate.connection.username", user);
-
-        properties.put("javax.persistence.jdbc.password", password);
-        properties.put("hibernate.connection.password", password);
-
-        HikariCPConnectionProvider cp = new HikariCPConnectionProvider();
-        cp.configure(properties);
-
-        this.connectionProviders.put(tenant, cp);
-    }
-
-    private ConnectionProvider getAnyConnectionProvider() {
-        return connectionProviders
-                .values().iterator().next();
-    }
-
-}
