+++ src/main/java/com/studies/ecommerce/models/BaseEntity.java
@@ -5,7 +5,6 @@ import lombok.Getter;
 import lombok.Setter;
 
 import javax.persistence.*;
-import javax.validation.constraints.NotBlank;
 
 @Getter
 @Setter
@@ -20,8 +19,5 @@ public class BaseEntity {
 
     @Version
     private Integer version;
-
-    @NotBlank
-    private String tenant;
 
 }
