--- src/main/java/com/studies/ecommerce/spring/config/JpaConfig.java
@@ -1,31 +1,0 @@
-package com.studies.ecommerce.spring.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.orm.jpa.JpaTransactionManager;
-import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
-import org.springframework.orm.jpa.vendor.HibernateJpaVendorAdapter;
-import org.springframework.transaction.annotation.EnableTransactionManagement;
-
-import javax.persistence.EntityManagerFactory;
-
-@EnableTransactionManagement
-@Configuration
-public class JpaConfig {
-
-    @Bean
-    public JpaTransactionManager jpaTransactionManager(EntityManagerFactory emf) {
-        return new JpaTransactionManager(emf);
-    }
-
-    @Bean
-    public LocalContainerEntityManagerFactoryBean entityManagerFactory() {
-        LocalContainerEntityManagerFactoryBean bean = new LocalContainerEntityManagerFactoryBean();
-
-        bean.setPersistenceUnitName("Ecommerce-PU");
-        bean.setJpaVendorAdapter(new HibernateJpaVendorAdapter());
-
-        return bean;
-    }
-
-}
