--- src/main/java/com/studies/ecommerce/spring/config/WebConfig.java
@@ -1,22 +1,0 @@
-package com.studies.ecommerce.spring.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@EnableWebMvc
-@Configuration
-public class WebConfig implements WebMvcConfigurer {
-
-    @Override
-    public void addResourceHandlers(ResourceHandlerRegistry registry) {
-        registry.addResourceHandler("/resources/**")
-                .addResourceLocations("/resources/");
-
-        registry.addResourceHandler("/webjars/**")
-                .addResourceLocations("/webjars/")
-                .resourceChain(true);
-    }
-
-}
