+++ src/main/java/com/studies/ecommerce/util/ExecuteDDL.java
@@ -10,9 +10,6 @@ public class ExecuteDDL {
     public static void main(String[] args) {
         Map<String, String> properties = new HashMap<>();
 
-        properties.put("javax.persistence.jdbc.url",
-                "jdbc:mysql://localhost/ecommerce?createDatabaseIfNotExist=true&useTimezone=true&serverTimezone=UTC");
-
         properties.put("javax.persistence.schema-generation.database.action",
                 "drop-and-create");
