+++ src/main/resources/META-INF/persistence.xml
@@ -14,18 +14,10 @@
 
         <properties>
             <property name="javax.persistence.jdbc.url"
-                      value="jdbc:mysql://localhost/db1_ecommerce?createDatabaseIfNotExist=true&amp;useTimezone=true&amp;serverTimezone=UTC" />
-            <property name="javax.persistence.jdbc.user" value="root" />
+                      value="jdbc:mysql://localhost/ecommerce?createDatabaseIfNotExist=true&amp;useTimezone=true&amp;serverTimezone=UTC" />
+            <property name="javax.persistence.jdbc.user" value="ecommerce_user" />
             <property name="javax.persistence.jdbc.password" value="123456" />
             <property name="javax.persistence.jdbc.driver" value="com.mysql.cj.jdbc.Driver" />
-
-            <!--<property name="hibernate.multiTenancy" value="SCHEMA" />
-            <property name="hibernate.tenant_identifier_resolver"
-                      value="com.studies.ecommerce.hibernate.EcmCurrentTenantIdentifierResolver"/>
-            <property name="hibernate.multi_tenant_connection_provider"
-                      value="com.studies.ecommerce.hibernate.EcmSchemaMultiTenantConnectionProvider"/>
-            <property name="hibernate.multi_tenant_connection_provider"
-                      value="com.studies.ecommerce.hibernate.EcmMachineMultiTenantConnectionProvider"/>-->
 
             <property name="hibernate.hikari.maximumPoolSize" value="8" />
             <property name="hibernate.connection.provider_class"
