--- src/main/webapp/WEB-INF/pages/index.html
@@ -1,20 +1,0 @@
-<!doctype html>
-<html lang="pt-br"
-      xmlns:th="http://www.thymeleaf.org"
-      xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
-      layout:decorator="~{layout/default}">
-<head>
-    <title>Welcome!</title>
-</head>
-<body>
-
-<th:block layout:fragment="content">
-    <section class="my-4">
-        <div class="container">
-            <h2>Welcome!</h2>
-        </div>
-    </section>
-</th:block>
-
-</body>
-</html>
