--- src/main/webapp/WEB-INF/pages/products/products-form.html
@@ -1,60 +1,0 @@
-<!doctype html>
-<html lang="pt-br"
-      xmlns:th="http://www.thymeleaf.org"
-      xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
-      layout:decorator="~{layout/default}">
-<head>
-    <title>Product Form</title>
-</head>
-<body>
-
-<th:block layout:fragment="content">
-    <section class="my-4">
-        <div class="container">
-            <h2>Product Form</h2>
-        </div>
-    </section>
-
-    <section>
-        <div class="container">
-            <form method="post" th:object="${product}"
-                  th:action="${#httpServletRequest.getServletPath()}">
-                <div class="alert alert-warning alert-dismissible fade show" role="alert"
-                     th:if="${#fields.hasAnyErrors()}">
-                    <button type="button" class="close" data-dismiss="alert" aria-label="Close">
-                        <span aria-hidden="true">&times;</span>
-                    </button>
-
-                    <div th:each="erro: ${#fields.detailedErrors()}">
-                        <span th:text="${erro.fieldName + ': ' + erro.message}">
-                            Validation Message.</span>
-                    </div>
-                </div>
-
-                <div class="alert alert-success alert-dismissible fade show" role="alert"
-                     th:if="${!#strings.isEmpty(message)}">
-                    <button type="button" class="close" data-dismiss="alert" aria-label="Close">
-                        <span aria-hidden="true">&times;</span>
-                    </button>
-
-                    <span th:text="${message}">Success Message.</span>
-                </div>
-
-                <div class="form-group"
-                     th:if="${product.id}">
-                    <label for="id">Id</label>
-                    <input type="text" class="form-control-plaintext" disabled th:field="*{id}"/>
-                </div>
-
-                <div class="form-group">
-                    <label for="name">Name</label>
-                    <input type="text" class="form-control" placeholder="Name"  th:field="*{name}"/>
-                </div>
-
-                <button type="submit" class="btn btn-primary">Save</button>
-            </form>
-        </div>
-    </section>
-</th:block>
-</body>
-</html>
