+++ src/test/java/com/studies/ecommerce/EntityManagerTest.java
@@ -1,13 +1,30 @@
 package com.studies.ecommerce;
 
 import org.junit.After;
+import org.junit.AfterClass;
 import org.junit.Before;
+import org.junit.BeforeClass;
 
 import javax.persistence.EntityManager;
+import javax.persistence.EntityManagerFactory;
+import javax.persistence.Persistence;
 
-public class EntityManagerTest extends EntityManagerFactoryTest {
+public class EntityManagerTest {
+
+    protected static EntityManagerFactory entityManagerFactory;
 
     protected EntityManager entityManager;
+
+    @BeforeClass
+    public static void setUpBeforeClass() {
+        entityManagerFactory = Persistence
+                .createEntityManagerFactory("Ecommerce-PU");
+    }
+
+    @AfterClass
+    public static void tearDownAfterClass() {
+        entityManagerFactory.close();
+    }
 
     @Before
     public void setUp() {
