+++ src/test/java/com/studies/ecommerce/concurrency/PessimisticLockTest.java
@@ -1,16 +1,44 @@
 package com.studies.ecommerce.concurrency;
 
-import com.studies.ecommerce.EntityManagerFactoryTest;
 import com.studies.ecommerce.models.Product;
+import org.junit.AfterClass;
 import org.junit.Assert;
+import org.junit.BeforeClass;
 import org.junit.Test;
 
 import javax.persistence.EntityManager;
+import javax.persistence.EntityManagerFactory;
 import javax.persistence.LockModeType;
+import javax.persistence.Persistence;
 import java.time.LocalDateTime;
 import java.util.List;
 
-public class PessimisticLockTest extends EntityManagerFactoryTest {
+public class PessimisticLockTest {
+
+    protected static EntityManagerFactory entityManagerFactory;
+
+    @BeforeClass
+    public static void setUpBeforeClass() {
+        entityManagerFactory = Persistence
+                .createEntityManagerFactory("Ecommerce-PU");
+    }
+
+    @AfterClass
+    public static void tearDownAfterClass() {
+        entityManagerFactory.close();
+    }
+
+    private static void log(Object obj, Object... args) {
+        System.out.println(
+                String.format("[LOG " + System.currentTimeMillis() + "] " + obj, args)
+        );
+    }
+
+    private static void myWaiting(int seconds) {
+        try {
+            Thread.sleep(seconds * 1000);
+        } catch (InterruptedException e) {}
+    }
 
     @Test
     public void typedQueryLock() {
